feat:시험기록 확인 날짜형식 개선 및 내부모달에도 반영
